drawing recognition result
also StrokePath now returns copies of Lists - preventing modification

+++ Scribble0.0.1/Assets/Scripts/TemplateLoader.cs
@@ -20,10 +20,10 @@ public struct Template
 public struct Point
 {
     [XmlAttribute("X")]
-    public int x;
+    public float x;
 
     [XmlAttribute("Y")]
-    public int y;
+    public float y;
 }
 
 public class TemplateLoader : MonoBehaviour
